restore uint32 for Decimals

+++ app/ante/testutils/testutils.go
@@ -109,7 +109,7 @@ func (suite *AnteTestSuite) SetupTest() {
 	configurator.ResetTestChainConfig()
 	err := configurator.
 		WithChainConfig(chainConfig).
-		WithEVMCoinInfo(denom, int64(decimals)).
+		WithEVMCoinInfo(denom, uint32(decimals)).
 		Configure()
 	suite.Require().NoError(err)
